cleanup and some conversions related to scalability

+++ src/main/java/com/starcases/prime/intfc/PrimeRefIntfc.java
@@ -36,5 +36,5 @@ public interface PrimeRefIntfc extends Serializable
 	 * @return int representing representing index in
 	 * overall list of primes.
 	 */
-	int getPrimeRefIdx();
+	long getPrimeRefIdx();
 }
